Separate Energy and Health
Added health_ to creature (with Get and Set health)

Modified Creature initialisation so health_ and energy_ are initially set to 100.

Modified SetEnergy and SetHealth so that the maximum Value is 100.

Added HealthToEnergy and EnergyToHealth

Added the threshold for health and energy conversion to the environment class.

modified death condition to depend on health instead of energy.