Separate Health and Energy
Added health_ to creature class (with Get and Set health functions)

Modified Creature initialisation so health_ and energy_ are initially set to 100.

Modified SetEnergy and SetHealth so that the maximum value is 100.

Added HealthToEnergy and EnergyToHealth functions to creature class:
Currently, they are both changed by +/- 5. If energy is below 0, you lose health until energy = 5

Added the threshold for health and energy conversion to the environment class.

modified death condition to depend on health instead of energy (checked at the end of UpdateEnergy).

All energy and health updates start from EnergyUpdate or Eats.